+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.6.0 - 2016-02-03
+
+### Added
+
+- [#18](https://github.com/zendframework/zend-crypt/pull/18) adds documentation,
+  and publishes it to https://zendframework.github.io/zend-crypt/
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#4](https://github.com/zendframework/zend-crypt/pull/4) replaces
+  the zend-servicemanager with container-interop, and refactors the
+  various plugin managers to implement that interface instead of extending the
+  `AbstractPluginManager`.
+
 ## 2.5.3 - TBD
 
 ### Added
@@ -101,3 +123,4 @@ All notable changes to this project will be documented in this file, in reverse 
   `Zend\Crypt\PublicKey\Rsa::MODE_AUTO`, if you were not using them previously.)
 
   We recommend re-encrypting any such values using the new defaults.
+>>>>>>> hotfix/5